iasd

+++ db/schema.rb
@@ -26,15 +26,15 @@ ActiveRecord::Schema.define(version: 20150809094013) do
     t.integer  "opr",        default: 0
     t.integer  "target",     default: 0
     t.integer  "act",        default: 0
-    t.integer  "percent",    default: 0
-    t.integer  "pph",        default: 0
+    t.float    "percent",    default: 0.0
+    t.float    "pph",        default: 0.0
     t.integer  "defect_int", default: 0
     t.integer  "defect_ext", default: 0
-    t.integer  "rft",        default: 0
+    t.float    "rft",        default: 0.0
     t.text     "remark"
     t.integer  "jam"
-    t.datetime "created_at",             null: false
-    t.datetime "updated_at",             null: false
+    t.datetime "created_at",               null: false
+    t.datetime "updated_at",               null: false
   end
 
   create_table "homes", force: :cascade do |t|
